nilfs2: scan dat entries at snapshot creation/deletion time
To accurately count the number of live blocks in a segment, it is
important to take snapshots into account, because snapshots can
protect reclaimable blocks from being cleaned.

This patch uses the previously reserved de_rsv (renamed to de_ss)
field of the nilfs_dat_entry structure to store one of the snapshots
the corresponding block belongs to. One block can belong to many
snapshots, but there is no need to know all of them. It suffices to
know, that it belongs to one or more snapshots.

If however the snapshot stored in the de_ss field is deleted,
we need to check if the block belongs to any other snapshot. When a
snapshot is created it is stored in a sorted linked list and when it
is deleted the previous and next elements of this list are known.
Because the list is sorted we only have to check the block against the
previous and the next element to determine if it is reclaimable or
not.

For example if the current snapshot (in de_ss) is being removed
and neither the previous nor the next snapshot is in the range of
de_start to de_end, then it is guaranteed that the block doesn't
belong to any other snapshot and is reclaimable. On the other hand
if lets say the previous snapshot is in the range of de_start to
de_end, simply set de_ss to the previous snapshot and the block is
not reclaimable.

If a block is reclaimable, de_ss is set to the special value
NILFS_ENTRY_DEC, indicating that the corresponding SUFILE entry was
decremented and does not need to be decremented again in the future.
If a block on the other hand is not reclaimable, de_ss is set either
to a snapshot number or to NILFS_ENTRY_INC.

To implement this every DAT entry is scanned at snapshot
creation/deletion time and updated if needed. To avoid too many
update operations only potentially reclaimable blocks are ever
updated. For example if there are some deleted files and the
checkpoint to which these files belong is turned into a snapshot,
then su_nlive_blks is incremented for these blocks, which reverses
the decrement that happen when the files were deleted. If after
some time this snapshot is deleted, su_nlive_blks is decremented
again to reverse the previous increment.

It is important to note, that the GC effectively resets the segment
usage information by collecting blocks from different locations and
moving them to a new segment. Therefore the flags NILFS_ENTRY_INC and
NILFS_ENTRY_DEC have to be updated accordingly.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>