nilfs2: add ioctl() to update snapshot flags of DAT-Entries
A normal GC operation entails the moving of blocks to a new segment,
whereby the block number in de_blocknr and the value of de_ss are
updated.

By using NILFS_IOCTL_SET_SUINFO the GC can set the su_nlive_blks field
to any value, without moving any blocks. While in this case the
de_blocknr does not have to be updated the de_ss field may still be
invalidated if it contains the NILFS_ENTRY_DEC flag. NILFS_ENTRY_DEC
means, that the corresponding value in su_nlive_blks for this entry
was already decremented and does not have to be decremented again. But
NILFS_IOCTL_SET_SUINFO effectivly incremented the value, so the flags
have to be updated to reflect this change.

To update these flags, this patch introduces a new ioctl
NILFS_IOCTL_SET_INC_FLAGS, which can be used to flip the de_ss value
from NILFS_ENTRY_DEC to NILFS_ENTRY_INC.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>