nilfs2: improve inode allocation algorithm
The current algorithm NILFS2 uses, does not use any information about
the previous allocation or the parent directory of the inode. It simply
searches for a free entry in the ifile, always starting from
position 0. This patch introduces an improved allocation scheme. There
are no changes to the on-disk-format necessary.

First of all the algorithm distinguishes between files and directories.

File inodes start their search at the location of their parent
directory, so that they will be allocated near their parent. If the file
inode ends up on the same block as its parent, the common task of
listing the directory contents (e.g.: "ls") will be faster. Although
there is no guarantee, that any subsequent blocks after that will end up
near the block with the directory on disk, there is still a possibility
for performance improvement, because it can be expected, that subsequent
blocks are read ahead. Also the cleaner will write out blocks in
the order of their file offset, so there is an increased likelihood that
those blocks can be read in faster.

Directory inodes are allocated aligned to block boundaries and with
BITS_PER_LONG empty slots following them. The empty slots increase the
likelihood, that files can be allocated in the same block as their
parent. Furthermore the alignment improves performance, because
unaligned locations do not have to be searched. The location of the last
successful allocation of a directory is stored in memory and used as a
starting point for the next allocation. This value is periodically reset
to 0 to allow the algorithm to find previously deleted slots.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>