nilfs2: add kmem_cache for SUFILE cache nodes
This patch adds a kmem_cache to efficiently allocate SUFILE cache nodes.
One cache node contains a certain number of unsigned 32 bit values and
either a list_head, to string a number of nodes together into a linked
list, or an rcu_head to be able to use the node with an rcu
callback.

These cache nodes can be used to cache small changes to the SUFILE and
apply them later at segment construction.

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>